backend/pyproject.toml
[project]
name = "support-triage"
version = "0.1.0"
description = "Support Ticket Triage System"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1,<5.2",
    "django-ninja>=1.3,<2.0",
    "psycopg[binary]>=3.2,<4.0",
    "dj-database-url>=2.3,<3.0",
    "django-dramatiq>=0.11,<1.0",
    "dramatiq[redis]>=1.17,<2.0",
    "structlog>=24.0,<25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-django>=4.9,<5.0",
    "factory-boy>=3.3,<4.0",
    "ruff>=0.8,<1.0",
    "mypy>=1.13,<2.0",
    "django-stubs>=5.1,<6.0",
    "pytest-cov>=6.0,<7.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = "3.12"
strict = true

[tool.django-stubs]
django_settings_module = "config.settings.dev"
